fix: output selection
